requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.7.2
attrs==21.2.0
backports.functools-lru-cache==1.6.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
configparser==5.0.2
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-appconf==1.0.4
django-compressor==2.4.1
django-debug-toolbar==3.2.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
enum34==1.1.10
gunicorn==20.1.0
heroku3==4.2.3
idna==3.2
iniconfig==1.1.1
isort==5.9.3
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.10.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-openid==2.2.5
python3-openid==3.2.0
pytz==2021.1
rcssmin==1.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
